chatbot/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
transformers==4.44.0
torch==2.4.0
